requirements.txt
Adafruit-Blinka==5.7.0
adafruit-circuitpython-ads1x15==2.2.3
adafruit-circuitpython-busdevice==5.0.1
Adafruit-PlatformDetect==2.21.0
Adafruit-PureIO==1.1.7
autopep8==1.5.4
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
paho-mqtt==1.5.1
pandas==1.1.4
Pillow==8.0.1
pycodestyle==2.6.0
pyftdi==0.52.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.4
pyusb==1.1.0
rpi-ws281x==4.2.5
RPi.GPIO==0.7.0
scipy==1.5.4
six==1.15.0
sysv-ipc==1.0.1
toml==0.10.2
